moving storage testing in fake
