+++ internal/storage/fake/register_bucket_tests.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package storagetesting
+package fake
 
 import (
 	"errors"
